Upates diagnostics - renamed module to comply with conventions - checks NEW settings covers all variables
