+++ services/web/server/src/simcore_service_webserver/application.py
@@ -13,7 +13,7 @@ from .activity import setup_activity
 from .catalog import setup_catalog
 from .computation import setup_computation
 from .db import setup_db
-from .diagnostics_plugin import setup_diagnostics
+from .diagnostics import setup_diagnostics
 from .director import setup_director
 from .email import setup_email
 from .groups import setup_groups
