+++ services/web/server/src/simcore_service_webserver/diagnostics.py
@@ -15,11 +15,13 @@ from .diagnostics_core import (
     kMAX_AVG_RESP_LATENCY,
     kMAX_TASK_DELAY,
     kPLUGIN_START_TIME,
-    kSTART_SENSING_DELAY_SECS
+    kSTART_SENSING_DELAY_SECS,
 )
 from .diagnostics_entrypoints import create_rest_routes
 from .diagnostics_monitoring import setup_monitoring
 from .rest import APP_OPENAPI_SPECS_KEY
+from .diagnostics_config import DiagnosticsSettings
+
 
 log = logging.getLogger(__name__)
 
@@ -40,7 +42,7 @@ def setup_diagnostics(
     start_sensing_delay: Optional[float] = 60
 ):
     # NOTE: keep all environs getters inside setup so they can be patched easier for testing
-
+    settings_kwargs = {}
 
     #
     # Any task blocked more than slow_duration_secs is logged as WARNING
@@ -48,6 +50,8 @@ def setup_diagnostics(
     #
     if slow_duration_secs is None:
         slow_duration_secs = float(os.environ.get("AIODEBUG_SLOW_DURATION_SECS", 0.3))
+    else:
+        settings_kwargs["slow_duration_secs"] = slow_duration_secs
 
     log.info("slow_duration_secs = %3.2f secs ", slow_duration_secs)
 
@@ -60,6 +64,8 @@ def setup_diagnostics(
             10 * slow_duration_secs,
             float(os.environ.get("DIAGNOSTICS_MAX_TASK_DELAY", 0)),
         )  # secs
+    else:
+        settings_kwargs["max_task_delay"] = max_task_delay
 
     app[kMAX_TASK_DELAY] = max_task_delay
     log.info("max_task_delay = %3.2f secs ", max_task_delay)
@@ -73,6 +79,8 @@ def setup_diagnostics(
         max_avg_response_latency = float(
             os.environ.get("DIAGNOSTICS_MAX_AVG_LATENCY", 3)
         )  # secs
+    else:
+        settings_kwargs["max_avg_response_latency"] = max_avg_response_latency
 
     app[kMAX_AVG_RESP_LATENCY] = max_avg_response_latency
     log.info("max_avg_response_latency = %3.2f secs ", max_avg_response_latency)
@@ -83,9 +91,18 @@ def setup_diagnostics(
     app[kSTART_SENSING_DELAY_SECS] = start_sensing_delay
     log.info("start_sensing_delay = %3.2f secs ", start_sensing_delay)
 
-    # -------
+    # ----------------------------------------------
+    # TODO: temporary, just to check compatibility between
+    # trafaret and pydantic schemas
+    cfg = DiagnosticsSettings(**settings_kwargs)
+    assert cfg.slow_duration_secs == slow_duration_secs
+    assert cfg.max_task_delay == max_task_delay
+    assert cfg.max_avg_response_latency == max_avg_response_latency
+    assert start_sensing_delay == start_sensing_delay
+    # ---------------------------------------------
 
-    #
+    # -----
+
     # TODO: redesign ... too convoluted!!
     registry = IncidentsRegistry(order_by=attrgetter("delay_secs"))
     app[kINCIDENTS_REGISTRY] = registry
